Fix Scala 2.13 Deprecation Warnings (#65)
* Fix deprecation warnings for Scala 2.13.10

- Updated percentage method to take `Int`s, as that's all it's ever called with

* Replaced deprecated symbol literal syntax

* Replaced the actor-based wsclient with the built-in Java http client

Solves deprecation warnings with actor materializer

+++ src/main/scala/pa/model.scala
@@ -401,10 +401,10 @@ case class PlayerStatsSummaryOffense(
   shotsOnTarget: Stat,
   throwIns: Stat
 ) {
-  private def percentage(n: Float, m: Float): Int = {
+  private def percentage(n: Int, m: Int): Int = {
     if (0 == n) 0
     else if (0 == n + m) 100
-    else Math.round((n / (n + m)) * 100)
+    else Math.round((n.toFloat / (n + m)) * 100)
   }
 
   val shotsOnTargetPercentage = new Stat(
